Rename sync to syncclient.
