+++ tox.ini
@@ -4,11 +4,11 @@ envlist = py27,py34,flake8
 [testenv]
 commands =
     python --version
-    py.test --cov-report term-missing --cov-fail-under 100 --cov sync {posargs}
+    py.test --cov-report term-missing --cov-fail-under 100 --cov syncclient {posargs}
 deps = -rdev-requirements.txt
 install_command = pip install --process-dependency-links --pre {opts} {packages}
 
 [testenv:flake8]
-commands = flake8 sync
+commands = flake8 syncclient
 deps =
     flake8
